+++ PathFinder/Source/RenderPipeline/VertexStorageLocation.hpp
@@ -7,10 +7,10 @@ namespace PathFinder
 
     struct VertexStorageLocation
     {
-        uint64_t VertexBufferOffset = 0;
-        uint64_t VertexCount = 0;
-        uint64_t IndexBufferOffset = 0;
-        uint64_t IndexCount = 0;
+        uint32_t VertexBufferOffset = 0;
+        uint32_t VertexCount = 0;
+        uint32_t IndexBufferOffset = 0;
+        uint32_t IndexCount = 0;
     };
 
 }
